Add Company service, controller, and update Company model
The update adds a service (`CompanyService`) for the operations related to the `Company` model. In addition, a new controller (`CompanyController`) is incorporated to manage HTTP requests related to the Companies. Furthermore, the `Company` model is updated to ensure `legalName` and `registrationCode` fields are not null, ensuring the mandatory nature of these attributes. The same changes have been synchronized to the database schema.