Add specific validation for Estonian Personal ID
A method was added to the IndividualService class to use a new utility class - IdValidationUtils. The utility class contains methods to validate Estonian Personal ID by checking its format, date of birth and checksum. An exception will be thrown if an invalid ID is detected during the creation or updating of an individual.