Update Individual service and repository methods
Changed the 'findByEventId' method to 'findByEvent_EventId' in both the IndividualService and IndividualRepository classes. This refactor aims to adhere to the proper naming convention when accessing nested properties in Spring JPA query methods -> the underscore (_) in findByEvent_EventId is used to navigate through the properties of the event object in the Individual entity.