Enforce "NOT NULL" constraints and refactor exception handling
The commit enforces "NOT NULL" constraints in the SQL schema for various tables preventing the entry of null values in necessary fields. It also refactors exception handling in the IndividualService class by throwing a new EventNotFoundException instead of IndividualNotFoundException when an event is not found. The exception is more accurately represented this way. Changes were done as well in EventService to revise the method used in finding associated companies.