Add participant count calculation to EventService
A new method, calculateTotalParticipantsForEvent, has been added to the EventService. This method calculates the total number of participants for a given event by combining counts from Individual and Company repositories. Also, a new endpoint was created in EventController to return this count. Changes have also been made to the repositories and the Company model to support this new feature.