Refactor IndividualController and update SQL schema
Removed unused import Individual from IndividualController. Also, updated the SQL schema syntax in schema_file-db.sql file by removing unnecessary spaces after NOT NULL declarations and updating the conditions for columns in the Individuals and Companies tables to cascade on delete. This will maintain database consistency by removing entries related to the deleted events. Additionally, the NumberOfParticipants column in the Companies table is set to NOT NULL for data integrity.